pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "matrix-risk-engine"
version = "0.1.0"
description = "Integrated platform for quantitative research, backtesting, and risk analytics"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Matrix Risk Engine Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    # Data Storage - ArcticDB for versioned time series
    # Pin to 4.x - API breaking changes between majors
    "arcticdb>=4.0.0,<5.0.0",

    # Data Processing - Pandas for DataFrames
    # Pin to 2.0.x - ArcticDB compatibility tested
    "pandas>=2.0.0,<2.2.0",

    # Numerical Computing - NumPy
    # Pin carefully - ORE-SWIG C++ bindings sensitive to array interface
    "numpy>=1.24.0,<2.0.0",

    # Backtesting - VectorBT for vectorized simulation
    # Pin to 0.25.x - package structure changes between versions
    "vectorbt>=0.25.3,<0.26.0",

    # Optimization - cvxpy for portfolio optimization
    # Pin to 1.4.x - solver API sensitive
    "cvxpy>=1.4.0,<1.6.0",

    # Performance Analytics - QuantStats for tearsheets
    "quantstats>=0.0.62",

    # Statistical Computing
    "scipy>=1.10.0,<1.13.0",

    # Configuration
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",

    # Type checking at runtime
    "typing-extensions>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pandas-stubs>=2.0.0",
    "types-PyYAML>=6.0.0",
]

ore = [
    # ORE-SWIG Python bindings - install separately
    # Requires C++ QuantLib and ORE libraries
    # See: https://github.com/OpenSourceRisk/ORE-SWIG
]

notebooks = [
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[project.scripts]
matrix-data = "src.cli.data_loader:main"
matrix-backtest = "src.cli.backtest_runner:main"
matrix-risk = "src.cli.risk_calculator:main"
matrix-optimize = "src.cli.optimizer_runner:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (may require external services)",
    "performance: Performance benchmarks",
    "slow: Slow tests (>10 seconds)",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
plugins = []

[[tool.mypy.overrides]]
module = [
    "arcticdb.*",
    "vectorbt.*",
    "quantstats.*",
    "cvxpy.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.isort]
known-first-party = ["src"]
